portfolio/scripts/data_collection/requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
apscheduler==3.10.4
plotly==5.18.0
seaborn==0.13.0
matplotlib==3.8.2
numpy==1.24.3
jupyter==1.0.0
notebook==7.0.6
scikit-learn==1.3.2
statsmodels==0.14.1
textblob==0.17.1
jinja2==3.1.2
markdown2==2.4.10
pyarrow==14.0.1  # For parquet support
fastparquet==2023.10.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2 